requirements.txt
asgiref==3.8.1
Django==3.2.12
django-environ==0.7.0
django-redis==5.4.0
gunicorn==20.1.0
pillow==10.3.0
pytz==2024.1
redis==5.0.3
sqlparse==0.4.4
tzdata==2024.1
